requirements.txt
asgiref==3.8.1
Django==4.2.16
django-appconf==1.0.6
django-compressor==4.5.1
django-sass-processor==1.4.1
libsass==0.23.0
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.2
tzdata==2024.2
django-autocomplete-light~=3.11.0
django-tinymce~=4.1.0